refactor: model
