+++ src/domain/article/index.ts
@@ -1,4 +1,4 @@
-export { default as Article } from './article';
+export { default as Article } from './model/article';
 export { articleQuerySchema } from './schema/articleQuerySchema';
 export type { ArticleQueryParams } from './schema/articleQuerySchema';
 export type { ArticleRepository } from './repository/articleRepository';
