+++ src/domain/article/mapper/articleMapper.ts
@@ -1,5 +1,5 @@
 import { Article as PrismaArticle } from '@prisma/client';
-import Article from '@/domain/article/article';
+import Article from '@/domain/article/model/article';
 
 export default function fromPrismaToArticle(prismaArticle: PrismaArticle): Article {
   return new Article(
