+++ src/domain/article/model/article.ts
